api/requirements.txt
absl-py==0.7.1
amqp==2.4.2
asgiref==3.1.2
astor==0.7.1
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
billiard==3.6.0.0
celery==4.3.0
channels==2.2.0
constantly==15.1.0
daphne==2.3.0
Django==3.0.7
django-filter==2.1.0
djangorestframework==3.11.0
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jaconv==0.2.4
kombu==4.5.0
Markdown==3.1
mock==3.0.4
numpy==1.16.3
pbr==5.2.0
protobuf==3.7.1
psycopg2==2.8.2
PyHamcrest==1.9.0
pytz==2019.1
redis==3.2.1
six==1.12.0
sqlparse==0.3.0
termcolor==1.1.0
Twisted==19.2.0
txaio==18.8.1
vine==1.3.0
Werkzeug==0.15.2
zope.interface==4.6.0